backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
feedparser==6.0.11
anthropic==0.34.2
openai>=1.40.0
apscheduler==3.10.4
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.12
